backend/tests: refactor backend tests and add new tests
Refactor backend tests, add some utils to make easier to create new
tests and add tests for `backend.AccountsByKeystore` and
`backend.AccountsTotalBalanceByKeystore`.
